pyproject.toml
[tool.poetry]
name = "devsync"
version = "0.1.0"
description = "A modern platform for developer collaboration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.1.0"
channels = "^4.0.0"
daphne = "^4.0.0"
openai = "^1.0.0"
python-dotenv = "^1.0.0"
celery = "^5.3.0"
redis = "^5.0.0"
psycopg2-binary = "^2.9.0"
requests = "^2.31.0"
python-dateutil = "^2.8.2"
markdown = "^3.5.0"
django-htmx = "^1.17.0"
whitenoise = "^6.5.0"
django-tailwind = "^3.6.0"
gunicorn = "^21.2.0"
sentry-sdk = "^1.30.0"
django-storages = "^1.14.0"
boto3 = "^1.28.0"
django-anymail = "^10.1"
django-maintenance-mode = "^0.18.0"
django-defender = "^0.9.7"
django-csp = "^3.7"
uvicorn = "^0.24.0"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.2.0"
django-extensions = "^3.2.0"
ipython = "^8.12.0"
pytest = "^7.3.1"
pytest-django = "^4.5.2"
coverage = "^7.2.7"
black = "^23.7.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
Sphinx = "^7.1.2"
sphinx-rtd-theme = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 